Exception Handling (implemented on ElectricMeter and Rates)

+++ src/main/java/com/utn/udeetpfinal/apirest/services/implementations/ElectricMeterServiceImpl.java
@@ -1,5 +1,6 @@
 package com.utn.udeetpfinal.apirest.services.implementations;
 
+import com.utn.udeetpfinal.apirest.exceptions.invalid.ElectricMeterInvalidException;
 import com.utn.udeetpfinal.apirest.models.ElectricMeter;
 import com.utn.udeetpfinal.apirest.repositories.IElectricMeterRepository;
 import com.utn.udeetpfinal.apirest.services.interfaces.IElectricMeterService;
@@ -29,18 +30,21 @@ public class ElectricMeterServiceImpl implements IElectricMeterService {
     }
 
     @Override
-    public Page<ElectricMeter> getAll(Integer page, Integer size, List<Order>orders) {
+    public Page<ElectricMeter> getAll(Integer page, Integer size, List<Order>orders) throws ElectricMeterInvalidException {
         Pageable pageable = PageRequest.of(page, size, Sort.by(orders));
+        if(electricMeterRepository.findAll(pageable)==null){
+            throw new ElectricMeterInvalidException("Can't get the Electric Meter List");
+        }
         return electricMeterRepository.findAll(pageable);
     }
 
     @Override
-    public ElectricMeter getElectricMeterById(Long electricMeterId) {
-        return electricMeterRepository.findById(electricMeterId).orElse(null);
+    public ElectricMeter getElectricMeterById(Long electricMeterId) throws ElectricMeterInvalidException{
+        return electricMeterRepository.findById(electricMeterId).orElseThrow(() -> new ElectricMeterInvalidException("Electric Meter with id "+electricMeterId+" doesn't exists"));
     }
 
     @Override
-    public void updateElectricMeter(Long electricMeterId, ElectricMeter newElectricMeter) {
+    public void updateElectricMeter(Long electricMeterId, ElectricMeter newElectricMeter) throws ElectricMeterInvalidException {
         ElectricMeter electricMeter = this.getElectricMeterById(electricMeterId);
         electricMeter.setBrand(newElectricMeter.getBrand());
         electricMeter.setId_serial(newElectricMeter.getId_serial());
@@ -49,7 +53,10 @@ public class ElectricMeterServiceImpl implements IElectricMeterService {
     }
 
     @Override
-    public void deleteElectricMeter(Long electricMeterId) {
+    public void deleteElectricMeter(Long electricMeterId) throws ElectricMeterInvalidException{
+        if(getElectricMeterById(electricMeterId) == null){
+            throw new ElectricMeterInvalidException("Electric Meter with id "+electricMeterId+" doesn't exists");
+        }
         electricMeterRepository.deleteById(electricMeterId);
     }
 }
