+++ src/main/java/com/utn/udeetpfinal/apirest/services/implementations/RateServiceImpl.java
@@ -1,5 +1,6 @@
 package com.utn.udeetpfinal.apirest.services.implementations;
 
+import com.utn.udeetpfinal.apirest.exceptions.invalid.RateInvalidException;
 import com.utn.udeetpfinal.apirest.models.Rate;
 import com.utn.udeetpfinal.apirest.repositories.IRateRepository;
 import com.utn.udeetpfinal.apirest.services.interfaces.IRateService;
@@ -28,24 +29,29 @@ public class RateServiceImpl implements IRateService {
         IRateRepository.save(rate);
     }
 
-    public Page<Rate> getRates(Integer page, Integer size, List<Order> orders){
+    public Page<Rate> getRates(Integer page, Integer size, List<Order> orders) throws RateInvalidException {
         Pageable pageable = PageRequest.of(page, size, Sort.by(orders));
-
+        if(IRateRepository.findAll((pageable)) == null){
+            throw new RateInvalidException("Can't get the Rate List");
+        }
         return IRateRepository.findAll(pageable);
     }
 
-    public Rate getRateById(Long rateId){
-        return IRateRepository.findById(rateId).orElse(null);
+    public Rate getRateById(Long rateId) throws RateInvalidException{
+        return IRateRepository.findById(rateId).orElseThrow(() -> new RateInvalidException("Rate with id "+rateId+" doesn't exists"));
     }
 
-    public void updateRate(Long rateId, Rate newRate) {
+    public void updateRate(Long rateId, Rate newRate) throws RateInvalidException{
         Rate rate = this.getRateById(rateId);
         rate.setPrice(newRate.getPrice());
         rate.setType(newRate.getType());
         this.addRate(rate);
     }
 
-    public void deleteRate(Long rateId) {
+    public void deleteRate(Long rateId) throws RateInvalidException {
+        if(getRateById(rateId) == null){
+            throw new RateInvalidException("Rate with id "+rateId+" doesn't exists");
+        }
         IRateRepository.deleteById(rateId);
     }
 }
